Formatting changed, parameter bug fixed

+++ homework/schedule/schedule.hpp
@@ -6,4 +6,4 @@
 
 void schedule(std::function<void()>func, std::chrono::seconds sec);
 void schedule(std::function<void(int)>func, std::chrono::seconds sec, int a);
-void schedule(std::function<void(std::string, double)>func, std::chrono::seconds sec, std::string str, int a);
+void schedule(std::function<void(std::string, double)>func, std::chrono::seconds sec, std::string str, double a);
